workers/facts-worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6
python-dotenv>=1.0.0
openai>=1.55.0
